Rename character-detail to home-character-detail
