+++ src/pages/home-character-detail/character.tsx
