+++ src/pages/home-character-detail/index.ts
