+++ src/pages/index.ts
@@ -1,5 +1,5 @@
 export * from "./home";
 export * from "./episodes";
 export * from "./location";
-export * from "./character-detail";
+export * from "./home-character-detail";
 export * from "./episodes-detail";
